Add a test that ensures unregistering is reasonably quick
There was a bug in io_uring that caused a 1 second delay on shutdown due
to how registered resources were cleaned up. This test exposes the bug,
and verifies that it does not reoccur. While time based tests have the
potential to be flaky, locally with the fix it completes in 5ms, and so
I do not expect this to be a practical problem.

Signed-off-by: Dylan Yudaken <dylany@fb.com>